Move timeToZero to utils.js
