+++ src/App.jsx
@@ -21,19 +21,7 @@ import { SelectValue } from "./components/ui/select";
 import { SelectContent } from "./components/ui/select";
 import { SelectItem } from "./components/ui/select";
 import P90xWorkouts from "./workouts/P90xWorkouts";
-
-export function timeToZero(oldDate) {
-  const newDate = new Date(oldDate)
-
-  // Set Time to 00:00:00
-  newDate.setHours(0)
-  newDate.setMinutes(0)
-  newDate.setSeconds(0)
-
-  // Technically ms still different, but doesn't show up in default string form
-
-  return newDate
-}
+import { timeToZero } from "./lib/utils";
 
 export default function App() {
   const [date, setDate] = useState(timeToZero(new Date()));
